feat(release-channel): render releases + invitations in unified inbox feed
